Refactor packages for cucumber
